+++ packages/react-moveable/src/react-moveable/ables/Draggable.tsx
@@ -74,6 +74,7 @@ export default {
             gesto,
         } = state;
 
+        console.log("has", gesto);
         if (gesto) {
             return false;
         }